requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pytesseract>=0.3.10
boto3>=1.34.0
spacy>=3.7.0
python-multipart>=0.0.6
httpx>=0.26.0
tenacity>=8.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dateutil>=2.8.0
